[outputbuffer] create base class for the generic part
